PSEC-1871 refactor code to fix merge conflicts
