+++ bitwarden_manager/aws_secretsmanager_client.py
