+++ bitwarden_manager/bitwarden_manager.py
